refactor(bulk-export): add markdown/ barrel index.ts and update imports (design alignment)
Creates the markdown module's public entry point (design.md § File Structure Plan),
exposes only createBulkExportMarkdownRenderer, and updates the export step and its
spec to import through the barrel instead of reaching into internal files.
